requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
celery==5.3.4
django-celery-results==2.5.1
django-celery-beat==2.6.0
sqlalchemy==2.0.23